Assignment06 (Ninth Commit)

+++ src/A6Q1.java
@@ -11,10 +11,10 @@ import java.util.Scanner;
  */
 public class A6Q1 {
 
-    public static double circleArea(int radius) {
-        // Finding the area of the cirle using user's input of radius
+    public static double circleArea(double radius) {
+        // Calculating the area of the cirle using user's input of radius
         double ans = ((radius * radius) * Math.PI);
-        // Returns the answers
+        // Returns the variable 'ans' of the method
         return ans;
     }
 
@@ -22,11 +22,13 @@ public class A6Q1 {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        // Save the user's input
+        // Create new scanner
         Scanner input = new Scanner(System.in);
-        // Ask the user's their circle's radius
+        // Ask the user their circle's radius
         System.out.println("Please enter you circle's radius");
-        int circleArea = input.nextInt();
+        // Save input of user to variable 'circleArea'
+        double circleArea = input.nextDouble();
+        // Connecting the method to rest of code
         double ans = circleArea(circleArea);
         // Output the user's answer (area of their circle)
         System.out.println("Your circle's area is " + ans);
